feat: middleware 설정 변경(쿠키유무에서 세션내 user으로 변경)
